// frontend/src/api/generated/models/ResultComparison.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CompetitorComparison } from './CompetitorComparison';
import type { RecordComparison } from './RecordComparison';
import type { TargetResult } from './TargetResult';
export type ResultComparison = {
    target_result?: TargetResult;
    records?: RecordComparison;
    competitor_comparison?: Array<CompetitorComparison>;
};